refactor(ios): extract UsernameEditView from SettingsView
Move the username display/inline-edit block out of SettingsView into a
new UsernameEditView. The extracted view owns its edit state internally
(editing flag, draft, saving, error/success messages) and takes the
current user, an AppViewModel reference, and an external `updating` flag
as props. Behavior, accessibility identifiers, and validation are
unchanged.

Closes #419
