+++ ios/StillPointApp/Views/SettingsView.swift
@@ -8,18 +8,13 @@ struct SettingsView: View {
     @State private var aphorismsEnabled: Bool = false
     @State private var isUpdating = false
     @State private var isUpdatingAphorisms = false
-    @State private var editingUsername = false
-    @State private var usernameDraft = ""
-    @State private var savingUsername = false
-    @State private var usernameFieldError: String?
-    @State private var usernameSuccessMessage: String?
     @State private var showDeleteAccountDialog = false
     @State private var showDeleteAccountConfirm = false
     @State private var isDeletingAccount = false
     @State private var deleteAccountError = ""
     @State private var showDeleteAccountError = false
 
-    private var isSavingSettings: Bool { isUpdating || isUpdatingAphorisms || savingUsername }
+    private var isSavingSettings: Bool { isUpdating || isUpdatingAphorisms }
 
     var body: some View {
         NavigationStack {
@@ -44,7 +39,7 @@ struct SettingsView: View {
                         .tracking(2)
 
                     if let user = appVM.currentUser {
-                        usernameSection(for: user)
+                        UsernameEditView(user: user, appVM: appVM, updating: isSavingSettings)
 
                         HStack {
                             Text("Email")
@@ -263,11 +258,6 @@ struct SettingsView: View {
         .onAppear {
             syncFromCurrentUser()
         }
-        .onChange(of: appVM.currentUser?.username) { _, _ in
-            if !editingUsername {
-                syncFromCurrentUser()
-            }
-        }
         .onChange(of: appVM.currentUser?.isPublic) { _, _ in
             guard appVM.currentUser != nil else { return }
             syncFromCurrentUser()
@@ -307,152 +297,9 @@ struct SettingsView: View {
         .accessibilityIdentifier("settings.notificationsLink")
     }
 
-    @ViewBuilder
-    private func usernameSection(for user: UserDTO) -> some View {
-        VStack(alignment: .leading, spacing: SPSpacing.s2) {
-            if editingUsername {
-                TextField("Username", text: $usernameDraft)
-                    .font(SPFont.mono(15))
-                    .foregroundStyle(Color(SPColor.fg))
-                    .textInputAutocapitalization(.never)
-                    .autocorrectionDisabled()
-                    .disabled(isSavingSettings)
-                    .accessibilityIdentifier("settings.usernameField")
-                    .onChange(of: usernameDraft) { _, newValue in
-                        if newValue.count > UsernameValidation.maxLength {
-                            usernameDraft = String(newValue.prefix(UsernameValidation.maxLength))
-                        }
-                    }
-
-                HStack(spacing: SPSpacing.s2) {
-                    Button {
-                        Task { @MainActor in
-                            await saveUsername(currentUsername: user.username)
-                        }
-                    } label: {
-                        Text(savingUsername ? "Saving…" : "Save")
-                            .font(SPFont.mono(11, weight: .medium))
-                            .tracking(2)
-                    }
-                    .buttonStyle(.bordered)
-                    .disabled(isSavingSettings)
-                    .accessibilityIdentifier("settings.usernameSaveButton")
-
-                    Button("Cancel") {
-                        cancelUsernameEdit(savedUsername: user.username)
-                    }
-                    .font(SPFont.mono(11, weight: .medium))
-                    .tracking(2)
-                    .buttonStyle(.bordered)
-                    .disabled(isSavingSettings)
-                    .accessibilityIdentifier("settings.usernameCancelButton")
-                }
-
-                if let usernameFieldError {
-                    Text(usernameFieldError)
-                        .font(SPFont.mono(11))
-                        .foregroundStyle(SPColor.dangerMuted)
-                        .accessibilityIdentifier("settings.usernameError")
-                }
-            } else {
-                HStack(alignment: .center, spacing: SPSpacing.s2) {
-                    Text("Username")
-                        .font(SPFont.mono(13))
-                        .foregroundStyle(Color(SPColor.fg3))
-                    Spacer(minLength: SPSpacing.s2)
-                    Text(user.username)
-                        .font(SPFont.mono(13))
-                        .foregroundStyle(Color(SPColor.fg))
-                        .lineLimit(1)
-                        .accessibilityIdentifier("settings.usernameDisplay")
-                    Button("Edit") {
-                        beginUsernameEdit(savedUsername: user.username)
-                    }
-                    .font(SPFont.mono(10, weight: .medium))
-                    .tracking(2)
-                    .buttonStyle(.bordered)
-                    .accessibilityLabel("Edit username")
-                    .accessibilityIdentifier("settings.usernameEditButton")
-                    .disabled(isSavingSettings)
-                }
-
-                if let usernameSuccessMessage {
-                    Text(usernameSuccessMessage)
-                        .font(SPFont.mono(11))
-                        .foregroundStyle(SPColor.green)
-                        .accessibilityIdentifier("settings.usernameSuccess")
-                }
-            }
-        }
-    }
-
     private func syncFromCurrentUser() {
         isPublic = appVM.currentUser?.isPublic ?? false
         aphorismsEnabled = appVM.currentUser?.aphorismsEnabled ?? false
-        if !editingUsername, let u = appVM.currentUser {
-            usernameDraft = u.username
-        }
-    }
-
-    private func beginUsernameEdit(savedUsername: String) {
-        usernameDraft = savedUsername
-        usernameFieldError = nil
-        usernameSuccessMessage = nil
-        editingUsername = true
-    }
-
-    private func cancelUsernameEdit(savedUsername: String) {
-        editingUsername = false
-        usernameFieldError = nil
-        usernameDraft = savedUsername
-    }
-
-    @MainActor
-    private func saveUsername(currentUsername: String) async {
-        guard !isSavingSettings else { return }
-
-        let trimmed = usernameDraft.trimmingCharacters(in: .whitespacesAndNewlines)
-        usernameFieldError = nil
-        usernameSuccessMessage = nil
-
-        if trimmed == currentUsername {
-            editingUsername = false
-            return
-        }
-
-        guard UsernameValidation.isValid(trimmed) else {
-            usernameFieldError = UsernameValidation.errorMessage
-            return
-        }
-
-        savingUsername = true
-        defer { savingUsername = false }
-
-        do {
-            let updated = try await APIClient.shared.updateSettings(username: trimmed)
-            appVM.applySettingsUser(updated)
-            usernameDraft = updated.username
-            editingUsername = false
-            usernameSuccessMessage = "Username updated"
-        } catch let error as APIError {
-            usernameFieldError = Self.message(for: error)
-        } catch {
-            usernameFieldError = "Could not update username. Please try again."
-        }
-    }
-
-    private static func message(for error: APIError) -> String {
-        switch error.status {
-        case 409:
-            return error.message
-        case 400:
-            if error.message == UsernameValidation.errorMessage {
-                return UsernameValidation.errorMessage
-            }
-            return error.message
-        default:
-            return "Could not update username. Please try again."
-        }
     }
 
     private var appVersionFooter: String {
